Move deployer and config tracker to canary package
