+++ pkg/canary/deployer_test.go
@@ -1,4 +1,4 @@
-package controller
+package canary
 
 import (
 	"testing"
@@ -8,8 +8,8 @@ import (
 )
 
 func TestCanaryDeployer_Sync(t *testing.T) {
-	mocks := SetupMocks(false)
-	err := mocks.deployer.Sync(mocks.canary)
+	mocks := SetupMocks()
+	err := mocks.deployer.Initialize(mocks.canary)
 	if err != nil {
 		t.Fatal(err.Error())
 	}
@@ -94,8 +94,8 @@ func TestCanaryDeployer_Sync(t *testing.T) {
 }
 
 func TestCanaryDeployer_IsNewSpec(t *testing.T) {
-	mocks := SetupMocks(false)
-	err := mocks.deployer.Sync(mocks.canary)
+	mocks := SetupMocks()
+	err := mocks.deployer.Initialize(mocks.canary)
 	if err != nil {
 		t.Fatal(err.Error())
 	}
@@ -106,7 +106,7 @@ func TestCanaryDeployer_IsNewSpec(t *testing.T) {
 		t.Fatal(err.Error())
 	}
 
-	isNew, err := mocks.deployer.IsNewSpec(mocks.canary)
+	isNew, err := mocks.deployer.HasDeploymentChanged(mocks.canary)
 	if err != nil {
 		t.Fatal(err.Error())
 	}
@@ -117,8 +117,8 @@ func TestCanaryDeployer_IsNewSpec(t *testing.T) {
 }
 
 func TestCanaryDeployer_Promote(t *testing.T) {
-	mocks := SetupMocks(false)
-	err := mocks.deployer.Sync(mocks.canary)
+	mocks := SetupMocks()
+	err := mocks.deployer.Initialize(mocks.canary)
 	if err != nil {
 		t.Fatal(err.Error())
 	}
@@ -162,8 +162,8 @@ func TestCanaryDeployer_Promote(t *testing.T) {
 }
 
 func TestCanaryDeployer_IsReady(t *testing.T) {
-	mocks := SetupMocks(false)
-	err := mocks.deployer.Sync(mocks.canary)
+	mocks := SetupMocks()
+	err := mocks.deployer.Initialize(mocks.canary)
 	if err != nil {
 		t.Error("Expected primary readiness check to fail")
 	}
@@ -180,8 +180,8 @@ func TestCanaryDeployer_IsReady(t *testing.T) {
 }
 
 func TestCanaryDeployer_SetFailedChecks(t *testing.T) {
-	mocks := SetupMocks(false)
-	err := mocks.deployer.Sync(mocks.canary)
+	mocks := SetupMocks()
+	err := mocks.deployer.Initialize(mocks.canary)
 	if err != nil {
 		t.Fatal(err.Error())
 	}
@@ -202,8 +202,8 @@ func TestCanaryDeployer_SetFailedChecks(t *testing.T) {
 }
 
 func TestCanaryDeployer_SetState(t *testing.T) {
-	mocks := SetupMocks(false)
-	err := mocks.deployer.Sync(mocks.canary)
+	mocks := SetupMocks()
+	err := mocks.deployer.Initialize(mocks.canary)
 	if err != nil {
 		t.Fatal(err.Error())
 	}
@@ -224,8 +224,8 @@ func TestCanaryDeployer_SetState(t *testing.T) {
 }
 
 func TestCanaryDeployer_SyncStatus(t *testing.T) {
-	mocks := SetupMocks(false)
-	err := mocks.deployer.Sync(mocks.canary)
+	mocks := SetupMocks()
+	err := mocks.deployer.Initialize(mocks.canary)
 	if err != nil {
 		t.Fatal(err.Error())
 	}
@@ -263,8 +263,8 @@ func TestCanaryDeployer_SyncStatus(t *testing.T) {
 }
 
 func TestCanaryDeployer_Scale(t *testing.T) {
-	mocks := SetupMocks(false)
-	err := mocks.deployer.Sync(mocks.canary)
+	mocks := SetupMocks()
+	err := mocks.deployer.Initialize(mocks.canary)
 	if err != nil {
 		t.Fatal(err.Error())
 	}
@@ -279,5 +279,4 @@ func TestCanaryDeployer_Scale(t *testing.T) {
 	if *c.Spec.Replicas != 2 {
 		t.Errorf("Got replicas %v wanted %v", *c.Spec.Replicas, 2)
 	}
-
 }
